Backend/requirements.old.txt
absl-py==1.0.0
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
amqp==5.0.9
asgi-redis==1.1.0
asgiref==3.5.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bleach==5.0.1
bokeh==2.4.3
cachetools==4.2.4
celery==5.2.3
Cerberus==1.3.4
certifi==2021.10.8
cffi==1.15.0
channels==3.0.5
channels-redis==3.4.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorcet==3.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cx-Oracle==7.3.0
daphne==3.0.2
debugpy==1.5.1
decorator==5.1.0
Deprecated==1.2.13
Django==3.0.1
django-4-jet @ https://github.com/Barukimang/django-jet/archive/dev.zip
django-activity-stream==1.4.1
django-axes==5.39.0
django-better-admin-arrayfield==1.4.2
django-bootstrap5==21.3
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-decouple==2.1
django-environ==0.9.0
django-filter==21.1
django-formtools==2.4
django-hstore==1.4.2
django-ipware==4.0.2
django-jalali==3.3.0
django-jsonfield-backport==1.0.5
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-object-permissions==1.4.6
django-otp==1.1.3
django-phonenumber-field==6.4.0
django-scheduler==0.9.5
django-softdelete==0.9.4
django-two-factor-auth==1.14.0
django-utils-six==2.0
django-webpush==0.3.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
drf-yasg==1.20.0
entrypoints==0.3
environ==1.0
Flask==2.0.2
flatbuffers==2.0
frozenlist==1.3.0
gast==0.4.0
geoip2==4.5.0
google-api-python-client==1.4.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
gunicorn==20.1.0
gviz-api==1.10.0
h5py==3.5.0
HeapDict==1.0.1
hiredis==2.0.0
holoviews==1.15.1
http-ece==1.1.0
httplib2==0.20.4
hvplot==0.8.1
hyperlink==21.0.0
icalendar==4.0.7
idna==3.3
incremental==21.3.0
inflection==0.5.1
influxdb==5.3.1
influxdb-client==1.25.0
Interface==2.11.1
ipykernel==6.5.0
ipython==7.29.0
itsdangerous==2.0.1
itypes==1.2.0
jdatetime==3.6.2
jedi==0.18.0
Jinja2==3.0.3
jsonfield==3.1.0
jupyter-client==7.0.6
jupyter-core==4.9.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kombu==5.2.3
libclang==12.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
MouseInfo==0.1.3
mqttasgi==1.1.2
msgpack==1.0.4
msgpack-python==0.5.6
multidict==6.0.2
mysqlclient==2.0.3
nest-asyncio==1.5.1
numpy==1.22.1
oauth2client==4.1.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.0
panel==0.14.0
param==1.12.2
parso==0.8.2
phonenumbers==8.13.0
pickleshare==0.7.5
Pillow==9.0.0
pipreqs==0.4.11
prompt-toolkit==3.0.24
protobuf==3.19.1
psycopg2==2.9.3
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pyct==0.4.8
PyGetWindow==0.0.9
Pygments==2.9.0
PyJWT==2.3.0
pymemcache==3.5.2
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.19.2
python-memcached==1.59
pytweening==1.0.4
pytz==2021.3
pyviz-comms==2.2.1
pywebpush==1.9.4
pywin32==302
PyYAML==6.0
pyzmq==22.3.0
qrcode==7.3.1
reactivex==4.0.4
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.9
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
Rx==3.2.0
schedule==1.1.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
streamz==0.6.4
suds-py3==1.4.0.0
swapper==1.3.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-profile==2.5.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.4.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
yarg==0.1.9
yarl==1.7.2
zict==2.2.0
zope.event==4.4
zope.interface==5.4.0
zope.schema==6.0.0
